requirements.txt
transformers==4.35.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
streamlit==1.28.0
torch==2.0.1
spacy==3.7.2
